WL-3465 WL-661 WL-3233 Removed old web-content tool.
The old web content tool came back in through a bad merge to Sakai because the web app load order is non-deterministic and the ToolManager lets another web app take over the registration for a tool.

We were getting some app-servers starting with the new web content tool and some with the old web content tool.